+++ geotrek/tourism/factories.py
@@ -13,15 +13,6 @@ from geotrek.trekking.factories import TrekFactory
 from django.conf import settings
 
 
-class DataSourceFactory(factory.Factory):
-    FACTORY_FOR = models.DataSource
-
-    title = factory.Sequence(lambda n: u"DataSource %s" % n)
-    url = factory.Sequence(lambda n: u"http://%s.com" % n)
-    type = models.DATA_SOURCE_TYPES.GEOJSON
-    pictogram = get_dummy_uploaded_image()
-
-
 class InformationDeskTypeFactory(factory.Factory):
     FACTORY_FOR = models.InformationDeskType