+++ geotrek/tourism/urls.py
@@ -12,8 +12,6 @@ from . import serializers as tourism_serializers
 
 urlpatterns = patterns(
     '',
-    url(r'^api/datasource/datasources.json$', tourism_views.DataSourceList.as_view(), name="datasource_list_json"),
-    url(r'^api/datasource/datasource-(?P<pk>\d+).geojson$', tourism_views.DataSourceGeoJSON.as_view(), name="datasource_geojson"),
     url(r'^api/(?P<lang>\w\w)/information_desks.geojson$', tourism_views.InformationDeskViewSet.as_view({'get': 'list'}), name="information_desk_geojson"),
     url(r'^api/(?P<lang>\w\w)/information_desks-(?P<type>\d+).geojson$', tourism_views.InformationDeskViewSet.as_view({'get': 'list'})),
     url(r'^api/treks/(?P<pk>\d+)/information_desks.geojson$', tourism_views.TrekInformationDeskViewSet.as_view({'get': 'list'}), name="trek_information_desk_geojson"),